requirements.txt
django-excel==0.0.10
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.6
qrcode==6.0
